refactor: Rename deciVoltage to batteryVoltageMilliVolts and update related methods
- Changed variable and method names from deciVoltage to batteryVoltageMilliVolts for clarity and consistency.
- Updated calculations to ensure battery voltage is handled in millivolts.
- Removed deprecated method for backward compatibility and adjusted telemetry data structures accordingly.
- Ensured all references to battery voltage across the codebase are consistent with the new naming convention.

+++ src/Telemetry/TelemetryData.h
@@ -12,7 +12,7 @@ struct TelemetryData {
     bool isReady;
 
     // Battery
-    uint32_t batteryVoltageMilliVolts;  // mV (ex: 44100 = 44.100V)
+    uint16_t batteryVoltageMilliVolts;  // mV (ex: 44100 = 44.100V, max 60V = 60000 mV)
     uint32_t batteryCurrentMilliAmps;    // mA (ex: 5000 = 5.000A)
 
     // Motor
@@ -30,7 +30,7 @@ struct TelemetryData {
  * Conversion functions for display purposes only
  * These return float only when necessary for formatting (e.g., in Xctod)
  */
-inline float milliVoltsToVolts(uint32_t mV) {
+inline float milliVoltsToVolts(uint16_t mV) {
     return mV / 1000.0f;
 }
